# source/swagger.yaml
openapi: 3.0.0
info:
  title: Akkor API
  description: API for Akkor hotel website.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /room/search:
    get:
      summary: Search for rooms based on query parameters
      description: Allows searching for rooms using different fields such as hotel_id, type, and capacity.
      operationId: getRoomByField
      tags:
        - Room
      parameters:
        - name: id
          in: query
          required: false
          description: ID of the room.
          schema:
            type: string
        - name: hotel_id
          in: query
          required: false
          description: ID of the hotel the room belongs to.
          schema:
            type: string
        - name: type_room
          in: query
          required: false
          description: Type of room (e.g., suite, deluxe).
          schema:
            type: string
        - name: max_nb_people
          in: query
          required: false
          description: Maximum number of people allowed in the room.
          schema:
            type: integer
        - name: number_of_room
          in: query
          required: false
          description: Total number of such rooms available.
          schema:
            type: integer
        - name: page
          in: query
          required: false
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of results per page.
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successfully retrieved rooms based on filters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: object
                    properties:
                      rooms:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            hotel_id:
                              type: integer
                            type_room:
                              type: string
                            max_nb_people:
                              type: integer
                            number_of_room:
                              type: integer
                            description:
                              type: string
                            createdAt:
                              type: string
                              format: date-time
                      totalCount:
                        type: integer
                      totalPages:
                        type: integer
                      currentPage:
                        type: integer
        '400':
          description: Invalid query parameter.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "One of the fields cannot be used."
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "An error has occurred."
...
  /room/{id}:
    put:
      summary: Update a room
      description: Allows admins to update room details. (Admin only)
      operationId: updateRoom
      tags:
        - Room
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the room to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotel_id:
                  type: integer
                  description: ID of the hotel the room belongs to.
                type_room:
                  type: string
                  description: Type of room.
                max_nb_people:
                  type: integer
                  description: Maximum number of people.
                number_of_room:
                  type: integer
                  description: Total number of such rooms.
                description:
                  type: string
                  description: Room description.
      responses:
        '200':
          description: Room updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: "Room updated."
        '404':
          description: Room not found.
        '500':
          description: Internal server error.

    delete:
      summary: Delete a room
      description: Allows admins to delete a room. (Admin only)
      operationId: deleteRoom
      tags:
        - Room
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the room to delete.
          schema:
            type: string
      responses:
        '200':
          description: Room deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: "Room deleted."
        '404':
          description: Room not found.
        '500':
          description: Internal server error.